requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
python-dotenv
graphene-django